backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
bson==0.5.10
python-dateutil==2.8.2
face-recognition==1.3.0
numpy==1.24.3 